django-skeleton/requirements.txt
aiohttp==3.8.3; python_version >= "3.6"
aiosignal==1.3.1; python_version >= "3.7"
appdirs==1.4.4
asgiref==3.5.2; python_version >= "3.8" and python_version < "4.0"
async-timeout==4.0.2; python_version >= "3.6"
attrs==22.1.0; python_version >= "3.6"
azure-common==1.1.28; python_version >= "3.7"
azure-core==1.26.1; python_version >= "3.7"
azure-servicebus==7.8.1; python_version >= "3.7"
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.8"
bcrypt==4.0.1; python_version >= "3.6"
boto3==1.24.69; python_version >= "3.7"
botocore==1.27.69; python_version >= "3.7"
cachetools==5.2.0; python_version >= "3.7" and python_version < "4.0"
certifi==2022.6.15; python_version >= "3.7" and python_version < "4"
cffi==1.15.1; python_version >= "3.6"
charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
confuse==2.0.0; python_version >= "3.6" and python_version < "4"
cryptography==38.0.3; python_version >= "3.6"
deprecated==1.2.13; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
django-cors-headers==3.13.0; python_version >= "3.7"
django-guid==3.3.0; python_version >= "3.7" and python_version < "4.0"
django-migration-linter==4.1.0
django-redis==5.2.0; python_version >= "3.6"
django==4.1; python_version >= "3.8"
djangorestframework-simplejwt==5.2.0; python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
dnspython==2.2.1; python_version >= "3.7" and python_version < "4.0"
frozenlist==1.3.3; python_version >= "3.7"
greenlet==1.1.3; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0"
gunicorn==20.1.0; python_version >= "3.5"
idna==3.3; python_version >= "3.7" and python_version < "4"
isodate==0.6.1; python_version >= "3.7"
jmespath==1.0.1; python_version >= "3.7"
jsonfield==3.1.0; python_version >= "3.6"
msrest==0.7.1; python_version >= "3.7"
multidict==6.0.2; python_version >= "3.7"
mysqlclient==2.1.1; python_version >= "3.5"
mysql-connector-python
numpy==1.23.3
oauthlib==3.2.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
orjson==3.8.2; python_version >= "3.7"
packaging==21.3; python_version >= "3.6"
pandas==1.4.4; python_version >= "3.8"
paramiko==2.12.0
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pydantic==1.10.1; python_version >= "3.7"
pyjwt==2.4.0; python_version >= "3.7"
pymongo==4.3.2; python_version >= "3.7"
pynacl==1.5.0; python_version >= "3.6"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.6"
python-amazon-sp-api==0.17.3
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
python-dotenv==0.20.0; python_version >= "3.5"
python-json-logger==2.0.4; python_version >= "3.5"
pytz==2022.2.1; python_version >= "3.8"
pyyaml==6.0; python_version >= "3.6" and python_version < "4"
redis==4.3.4; python_version >= "3.6"
requests-oauthlib==1.3.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
requests==2.28.1; python_version >= "3.7" and python_version < "4"
s3transfer==0.6.0; python_version >= "3.7"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
sqlalchemy==1.4.41; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
sqlparse==0.4.2; python_version >= "3.8" and python_version < "4.0"
sshtunnel==0.4.0
toml==0.10.2; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
typing-extensions==4.3.0; python_version >= "3.7"
tzdata==2022.2; sys_platform == "win32" and python_version >= "3.8" and python_version < "4.0"
uamqp==1.6.3; python_version >= "3.7"
urllib3==1.26.12; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
whitenoise==6.2.0; python_version >= "3.7"
wrapt==1.14.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
yarl==1.8.1; python_version >= "3.7"
